refactor: modularize codebase and remove dead code
- Move llm-client.factory.ts and web-scraper.ts from benchmark/utils/ to
  src/utils/ since they are shared by runtime services
- Split agent-tools.ts (607 LoC) into src/services/tools/ with separate
  modules for definitions, search, analysis, knowledge, and context handlers
- Split llm.service.ts (360 LoC) into src/services/llm/ with separate
  modules for query refinement, paper analysis, and literature review
- Remove dead exports: getPaperDetails, getPaperByExternalId,
  extractFromPdfUrl, extractFromHtml, getAvailableConfigs, scrapeUrls,
  scrapeArxiv, scrapeDoi, isValidUrl, isResearchPaperUrl, extractDomain
- Remove unused SemanticScholarPaper/SearchResponse from database.models
  (moved inline to semantic-scholar.provider.ts)
- Fix AuthenticatedRequest duplication: controller now imports from middleware
- Fix middleware filename typo: auth.middlware.ts -> auth.middleware.ts
