+++ src/benchmark/runners/generation-bench.ts
@@ -4,7 +4,7 @@
  */
 
 import { searchService } from '../../services/search.service';
-import { llmService } from '../../services/llm.service';
+import { llmService } from '../../services/llm';
 import {
     extractCitations,
     calculateCitationDensity,
