+++ src/services/tools/index.ts
@@ -1,0 +1,2 @@
+export { AGENT_TOOLS } from './tool-definitions';
+export { toolExecutor } from './tool-executor';
